Hide HTTPClient implementation behind a protocol
- added AWSHTTPClient protocol, generic request and response structs
- Use generic request and response structures throughout code
- Renamed internal HTTPClient to NIOTSHTTPClient in preparation for making it NIO TS only
- Have both NIOTSHTTPClient and AsyncHTTPClient inherit from AWSHTTPClient
- Added createHTTPClient() which will create a client based on whether we are NIO TS or not
