+++ Sources/AWSSDKSwiftCore/Message/AWSResponse.swift
@@ -26,13 +26,13 @@ public struct AWSResponse {
     ///     - from: Raw HTTP Response
     ///     - serviceProtocol: protocol of service (.json, .xml, .query etc)
     ///     - raw: Whether Body should be treated as raw data
-    init(from response: HTTPResponseDescription, serviceProtocolType: ServiceProtocolType, raw: Bool = false) throws {
+    init(from response: AWSHTTPResponse, serviceProtocolType: ServiceProtocolType, raw: Bool = false) throws {
         self.status = response.status
         self.headers = AWSResponse.createHeaders(from: response)
         self.body = try AWSResponse.createBody(from: response, serviceProtocolType: serviceProtocolType, raw: raw)
     }
 
-    private static func createHeaders(from response: HTTPResponseDescription) -> [String: String] {
+    private static func createHeaders(from response: AWSHTTPResponse) -> [String: String] {
         var responseHeaders: [String: String] = [:]
         for (key, value) in response.headers {
             responseHeaders[key.description] = value
@@ -41,10 +41,10 @@ public struct AWSResponse {
         return responseHeaders
     }
     
-    private static func createBody(from response: HTTPResponseDescription, serviceProtocolType: ServiceProtocolType, raw: Bool) throws -> Body {
+    private static func createBody(from response: AWSHTTPResponse, serviceProtocolType: ServiceProtocolType, raw: Bool) throws -> Body {
         var responseBody: Body = .empty
         
-        guard let data = response.bodyData, !data.isEmpty else  {
+        guard let data = response.body, !data.isEmpty else  {
             return .empty
         }
